Prevent the script from removing the generated ".pdf" file

+++ pypdfplot/functions.py
@@ -214,8 +214,9 @@ def finalize_pypdf(pw,
 
     ## Remove the generating python file or create a new purely Python one:
     if cleanup or not _pure_py:
-        if remove_file(_pypdf_fname, verbose=verbose):
-            warnings.warn(_pypdf_fname + ' removed:\nSaving script in editor will make it reappear...!\n')
+        if os.path.splitext(_pypdf_fname)[1] == '.py':
+            if remove_file(_pypdf_fname, verbose=verbose):
+                warnings.warn(_pypdf_fname + ' removed:\nSaving script in editor will make it reappear...!\n')
 
     ## Write the Python file if needed, and remove it if not:
     if cleanup:
